app/ai/gemini.py
import google.generativeai as genai
from app.ai.base import AIPlatform


class Gemini(AIPlatform):
    def __init__(self, api_key: str, system_prompt: str = None):
        self.api_key = api_key
        self.system_prompt = system_prompt
        genai.configure(api_key=api_key)

        # See more models here: https://developers.generativeai.google/models
        self.model = genai.GenerativeModel("gemini-2.5-flash-preview-09-2025")

    def chat(self, prompt: str) -> str:
        if self.system_prompt:
            prompt = f"{self.system_prompt}\n\n{prompt}"

        response = self.model.generate_content(prompt)
        return response.text
