pyproject.toml
[project]
name = "fast-api-for-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "google-generativeai>=0.8.5",
    "python-jose[cryptography]>=3.5.0",
    "requests>=2.32.3",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
